# pkg(ticdc): add a faster raw string rendering version of GenUpdateSQL