<!--
Thank you for contributing to TiFlow! 
Please read MD's [CONTRIBUTING](https://github.com/pingcap/tiflow/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve?
<!--
Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.
 -->

Issue Number: ref #8057

Task-2

### What is changed and how it works?

- Use raw string format to generate update sql.
- Note this PR doesn't replace `GenUpdate` with `GenUpdateFast`, we will use `GenUpdateFast` after more correctness check and test.
- Benchmark result is as following, **the CPU costs decreases to `10%-16%`, memory allocation decreases to `5%-11%`.** (Horizontal axis denotes the batch size of `RowChange`)

CPU time             |  memory allocation
:-------------------------:|:-------------------------:
![](https://user-images.githubusercontent.com/1527315/212044190-636201e1-516a-4207-ba79-2711276da900.png) | ![](https://user-images.githubusercontent.com/1527315/212044197-892c54df-d0b2-4bf8-8d6a-4a200f2efd1b.png)


```go
➜  go test -run='^$' -benchmem -bench '^(BenchmarkGenUpdate)$' github.com/pingcap/tiflow/pkg/sqlmodel
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tiflow/pkg/sqlmodel
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch1-8              162788              7273 ns/op            8145 B/op        167 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch2-8              114633             10409 ns/op           10034 B/op        227 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch4-8               76413             15544 ns/op           13747 B/op        346 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch8-8               46393             25777 ns/op           20724 B/op        583 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch16-8              25616             46324 ns/op           33784 B/op       1056 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch32-8              13632             87381 ns/op           62977 B/op       2002 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch64-8               6699            169962 ns/op          128919 B/op       3893 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQL-Batch128-8              3602            331583 ns/op          246086 B/op       7671 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch1-8          914386              1224 ns/op            1712 B/op         20 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch2-8          789146              1451 ns/op            2168 B/op         23 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch4-8          615506              1810 ns/op            3160 B/op         29 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch8-8          439963              2655 ns/op            5056 B/op         41 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch16-8         229112              5016 ns/op           12570 B/op         67 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch32-8         123585              9502 ns/op           27069 B/op        117 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch64-8          61982             19472 ns/op           63620 B/op        216 allocs/op
BenchmarkGenUpdate/OneColumnPK-GenUpdateSQLFast-Batch128-8         32689             36418 ns/op          122018 B/op        410 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch1-8           159885              7308 ns/op            8145 B/op        167 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch2-8           114904             10370 ns/op           10034 B/op        227 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch4-8            77086             15542 ns/op           13747 B/op        346 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch8-8            46633             25813 ns/op           20724 B/op        583 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch16-8           25890             46306 ns/op           33784 B/op       1056 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch32-8           13792             87409 ns/op           62976 B/op       2002 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch64-8            6789            169864 ns/op          128918 B/op       3893 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQL-Batch128-8           3518            340506 ns/op          246086 B/op       7671 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch1-8       958459              1223 ns/op            1712 B/op         20 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch2-8       781336              1457 ns/op            2168 B/op         23 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch4-8       653800              1812 ns/op            3160 B/op         29 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch8-8       447306              2660 ns/op            5056 B/op         41 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch16-8      236116              5015 ns/op           12570 B/op         67 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch32-8      125460              9503 ns/op           27069 B/op        117 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch64-8       62330             19395 ns/op           63620 B/op        216 allocs/op
BenchmarkGenUpdate/MultiColumnsPK-GenUpdateSQLFast-Batch128-8              32689             36641 ns/op          122018 B/op        410 allocs/op
PASS
ok      github.com/pingcap/tiflow/pkg/sqlmodel  44.062s
```


### Check List <!--REMOVE the items that are not applicable-->

#### Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test

#### Questions <!-- Authors should answer these questions and reviewers should consider these questions. -->

##### Will it cause performance regression or break compatibility?

##### Do you need to update user documentation, design documentation or monitoring documentation?

### Release note <!-- bugfixes or new features need a release note -->

```release-note
None
```